REFACTOR: Move more of Excel reader from `ext` to `src`.
