+++ Project.toml
@@ -23,7 +23,7 @@ XLSX = "fdbf4ff8-1666-58a4-91e7-1b58723a45e0"
 nauty_jll = "55c6dc9b-343a-50ca-8ff2-b71adb3733d5"
 
 [extensions]
-ExcelACSets = "XLSX"
+ExcelACSetsExt = "XLSX"
 NautyACSetsExt = "nauty_jll"
 
 [compat]
